Experimental (WAY too fast) Atavus rewards
I'd like to compare the gift he's giving with what the player has
equipped in the same slot, so that he'll only give the gift if it's
"better" by some metric or another.  At *least*, he shouldn't give out
equipment that they players already wears, but much better if it's
smarter than just that.